+ sửa 1 số lỗi chính tả

+++ FinalProject/View/Controller/Favorite/FavoriteViewController.swift
@@ -6,7 +6,7 @@ final class FavoriteViewController: ViewController {
 	@IBOutlet weak var tableView: UITableView!
 
 	// MARK: - Properties
-	var viewModel: FavoriteViewControllerModel!
+	var viewModel: FavoriteViewControllerModel?
 
 	// MARK: - Life Cycle
 	override func viewWillAppear(_ animated: Bool) {
@@ -35,6 +35,7 @@ final class FavoriteViewController: ViewController {
 
 	// MARK: - Action
 	@objc func deleteAllDataRealm() {
+		guard let viewModel = viewModel else { return }
 		viewModel.deleteAllDataRealm { (success) in
 			if success {
 				tableView.reloadData()
@@ -48,12 +49,14 @@ final class FavoriteViewController: ViewController {
 // MARK: - UITableViewDataSource
 extension FavoriteViewController: UITableViewDataSource {
 	func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
+		guard let viewModel = viewModel else { return 0 }
 		return viewModel.listItemFavorite?.count ?? 0
 	}
 
 	func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-		guard let cell = tableView.dequeueReusableCell(withIdentifier: Config.favoriteTableViewCell, for: indexPath) as? FavoriteTableViewCell else {
-			fatalError("Imposible case")
+		guard let cell = tableView.dequeueReusableCell(withIdentifier: Config.favoriteTableViewCell, for: indexPath) as? FavoriteTableViewCell,
+			let viewModel = viewModel else {
+				fatalError("Imposible case")
 		}
 		cell.viewModel = viewModel.getCellViewModel(at: indexPath)
 		return cell
@@ -68,6 +71,7 @@ extension FavoriteViewController: UITableViewDelegate {
 	func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
 		if editingStyle == .delete {
 			tableView.beginUpdates()
+			guard let viewModel = viewModel else { return }
 			viewModel.deleteObjectRealm(at: indexPath.row) { (success) in
 				if success {
 					tableView.deleteRows(at: [indexPath], with: .fade)
